Srm tender

+++ app/Http/Controllers/API/TenderMasterAPIController.php
@@ -605,6 +605,7 @@ WHERE
             return ['status' => false, 'message' => "Atleast ".$tenderMaster->min_approval_bid_opening." employee should selected"];
         }
 
+
         if($input['addCalendarDates'] === 0){
             $resValidate = $this->validateTenderHeader($input);
             if (!$resValidate['status']) {
@@ -684,10 +685,10 @@ WHERE
             return ['success' => false, 'message' => 'Document sales to time is required'];
         }
 
+        
         if ($document_sales_start_date > $document_sales_end_date) {
             return ['success' => false, 'message' => 'From date and time cannot be greater than the To date and time  for Document Sales'];
         }
-
 
         if(!isset(($input['pre_bid_clarification_start_time']))) {
             return ['success' => false, 'message' => 'Pre bid clarification from time is required'];
@@ -698,9 +699,23 @@ WHERE
 
         }
 
-
         if ($pre_bid_clarification_start_date > $pre_bid_clarification_end_date) {
             return ['success' => false, 'message' => 'From date and time cannot be greater than the To date and time  for Pre-bid Clarification'];
+        }
+
+
+        if(!isset(($input['site_visit_start_time']))) {
+            return ['success' => false, 'message' => 'Site visit from time is required'];
+
+        }
+
+        if(!isset(($input['site_visit_end_time']))) {
+            return ['success' => false, 'message' => 'Site visit to time is required'];
+
+        }
+
+        if ($site_visit_date > $site_visit_end_date) {
+            return ['success' => false, 'message' => 'From date and time cannot be greater than the To date and time  for Site Visit'];
         }
 
         if(!isset(($input['bid_submission_opening_time']))) {
@@ -717,20 +732,6 @@ WHERE
             return ['success' => false, 'message' => 'From date and time cannot be greater than the To date and time  for Bid Submission'];
         }
 
-        if(!isset(($input['site_visit_start_time']))) {
-            return ['success' => false, 'message' => 'Site visit from time is required'];
-
-        }
-
-        if(!isset(($input['site_visit_end_time']))) {
-            return ['success' => false, 'message' => 'Site visit to time is required'];
-
-        }
-
-
-        if ($site_visit_date > $site_visit_end_date) {
-            return ['success' => false, 'message' => 'From date and time cannot be greater than the To date and time  for Site Visit'];
-        }
 
 
         if(isset($document_sales_start_date) && $document_sales_start_date < $currenctDate || isset($bid_submission_opening_date) && $bid_submission_opening_date < $currenctDate ||isset($pre_bid_clarification_start_date) && $pre_bid_clarification_start_date < $currenctDate ||isset($site_visit_date) && $site_visit_date < $currenctDate) {
@@ -1149,7 +1150,10 @@ WHERE
         $currenctDate = Carbon::now();
         if (isset($input['calendarDates'])) {
             if (count($input['calendarDates']) > 0) {
-                CalendarDatesDetail::where('tender_id', $input['id'])->where('company_id', $input['company_id'])->delete();
+                if(isset($input['time_changed'])){
+                    CalendarDatesDetail::where('tender_id', $input['id'])->where('company_id', $input['company_id'])->delete();
+                }
+                
                 foreach ($input['calendarDates'] as $calDate) {
                     $fromTime =($calDate['from_time']) ? new Carbon($calDate['from_time']) : null;
                     $toTime = ($calDate['to_time']) ? new Carbon($calDate['to_time']) : null;
@@ -1980,7 +1984,6 @@ WHERE
             $data['from_date'] = $frm_date;
         }
 
-        
 
         if (isset($request['to_date'])) {
             $to_date = new Carbon($request['to_date']);
@@ -1997,20 +2000,19 @@ WHERE
 
         if (!empty($frm_date) && !empty($to_date)) {
             if ($frm_date > $to_date) {
-                return ['success' => false, 'message' => 'From date cannot be greater than the To date'];
+                return ['success' => false, 'message' => 'From date and time cannot be greater than the To date and time'];
             }
         }
-
-
 
         $data['updated_at'] = Carbon::now();
         $data['updated_by'] = $employee->employeeSystemID;
        
         DB::beginTransaction();
         try {
+
             $calendarDatesDetail = CalendarDatesDetail::where('calendar_date_id', $request['calenderDateTypeId'])
-                ->where('tender_id', $request['tenderMasterId'])
-                ->first();
+            ->where('tender_id', $request['tenderMasterId'])
+            ->first();
 
             if (empty($calendarDatesDetail)) {
                 return $this->sendError('Calendar Date Type not found');
